+++ chrome/background.js
@@ -212,6 +212,7 @@ chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
 chrome.runtime.onInstalled.addListener(() => {
   // Set defaults on installation
   chrome.storage.sync.set({ canaryToken: "xnlreveal" });
+  chrome.storage.sync.set({ showAlerts: true });
   chrome.storage.sync.set({ copyToClipboard: false });
   chrome.storage.sync.set({ checkDelay: "2" });
   chrome.storage.sync.set({ waybackRegex: "" });