+++ firefox/background.js
@@ -210,6 +210,7 @@ browser.runtime.onMessage.addListener(function (request, sender, sendResponse) {
 browser.runtime.onInstalled.addListener(() => {
   // Set defaults on installation
   browser.storage.sync.set({ canaryToken: "xnlreveal" });
+  browser.storage.sync.set({ showAlerts: true });
   browser.storage.sync.set({ copyToClipboard: false });
   browser.storage.sync.set({ checkDelay: "2" });
   browser.storage.sync.set({ waybackRegex: "" });